Use fzf-tmux if SESS_USE_FZF_TMUS is true
tmux is already being used, might as well be able to use the `fzf-tmux`
script to load fzf in a tmux pane. This also checks if `fzf-tmux`
exists, and recommends updating fzf if it doesn't. It's been part of fzf
since 2015, so that _shouldn't_ be to big of an ask.